DD-1965 Add packaging format extension (#11)
When creating a storage root, copy the packaging-format-registry and property-registry extensions to its extensions directory.
When adding a new version of an object, update the object-version-properties with a hard-coded (for now) packaging-format = DANS RDA BagPack Profile/0.1.0 property.
Upgraded dans-ocfl-java-extensions and dans-layer-store-lib dependencies, so that the files in the extension folder are now also backed by the database and can therefore be retrieved faster. This is necessary for updating object-version-properties to be feasible.